Changed project to .Lib - Read long desc
Updated the Starlight64 project to be a .lib instead of a .exe, this will allow users to instead build ontop of their engine instead of inside of the engine. This is the intended use of the Starlight64 engine HOWEVER this will not be pushed to the main development branch until I figure out why debug mode doesnt work.